+++ apex_legends_api/al_domain.py
@@ -23,7 +23,7 @@ class GameEvent(Event):
         self.game_length: int = event_dict.get('gameLength')
         self.legend_played: str = event_dict.get('legendPlayed')
         self.rank_score_change: str = event_dict.get('rankScoreChange')
-        self.game_data_trackers: list[DataTrackers] = list()
+        self.game_data_trackers: list = list()
         for tracker in event_dict.get('event'):
             self.game_data_trackers.append(DataTrackers(tracker))
 
@@ -47,12 +47,20 @@ class LevelEvent(Event):
         self.new_level = event_dict.get('event').get('newLevel')
 
 
+class RankEvent(Event):
+    """ event subclass for 'rank' events (ranked progression)"""
+    def __init__(self, event_dict: dict):
+        super().__init__(event_dict)
+        self.new_rank = event_dict.get('event').get('newRank')
+
+
 def event_factory(event_dict: dict) -> Event:
     """ a factory method for the different event types"""
     event_classes = {
         "Session": SessionEvent,
         "Level": LevelEvent,
-        "Game": GameEvent
+        "Game": GameEvent,
+        "Rank": RankEvent
     }
     return event_classes[event_dict.get('eventType')](event_dict)
 
@@ -189,10 +197,10 @@ class ALPlayer:
         self.selected_legend: SelectedLegend = SelectedLegend(
             selected_legend_dict=basic_player_stats_data['legends']['selected']
         )
-        self.all_legends: list[Legend] = list()
+        self.all_legends: list = list()
         for legend_name, legend_dict in basic_player_stats_data['legends']['all'].items():
             self.all_legends.append(Legend(legend_name=legend_name, legend_dict=legend_dict))
-        self.match_history = list()
+        self.match_history: list = list()
         if match_history:
             for event in match_history:
                 event_result = event_factory(event_dict=event)